By splitting `repair/row_level.cc` I save 10 minutes of build time for `$ ninja build/release/scylla`.

4.5.dev-0.20210406.ef4ad363c (after)
--

```
$ ninja build/release/scylla  13079.35s user 386.95s system 2569% cpu 8:44.12 total
$ ninja build/release/scylla  13574.93s user 404.36s system 2590% cpu 8:59.73 total
```

tail distribution of the slowest files after this branch on release in ms:

```
90488 build/release/repair/repair_meta.json
94184 build/release/distributed_loader.json
94890 build/release/api/column_family.json
97226 build/release/db/config.json
99380 build/release/cql3/query_processor.json
107981 build/release/alternator/executor.json
109456 build/release/transport/server.json
112378 build/release/db/system_keyspace.json
112606 build/release/db/commitlog/commitlog.json
116161 build/release/db/view/view.json
122211 build/release/gms/gossiper.json
123754 build/release/thrift/handler.json
125239 build/release/types.json
151471 build/release/db/schema_tables.json
153972 build/release/main.json
154930 build/release/service/storage_service.json
159369 build/release/database.json
166506 build/release/table.json
168228 build/release/cql3/statements/select_statement.json
169701 build/release/mutation_partition.json
205628 build/release/api/storage_service.json
262423 build/release/sstables/sstables.json
281853 build/release/service/storage_proxy.json
287454 build/release/message/messaging_service.json
320895 build/release/repair/repair.json
``` 

4.5.dev-0.20210406.c1badc631 (before)
--

```
$ ninja build/release/scylla  14094.45s user 408.50s system 1364% cpu 17:43.09 total
$ ninja build/release/scylla  13662.06s user 389.78s system 1370% cpu 17:05.43 total
```

tail distribution of the slowest files before this branch on release in ms:

```
88697 build/release/api/column_family.json
98477 build/release/db/config.json
101052 build/release/cql3/query_processor.json
108475 build/release/db/commitlog/commitlog.json
108559 build/release/transport/server.json
110517 build/release/db/view/view.json
111171 build/release/gms/gossiper.json
113413 build/release/db/system_keyspace.json
123365 build/release/types.json
123896 build/release/thrift/handler.json
138468 build/release/db/schema_tables.json
141379 build/release/service/storage_service.json
143828 build/release/main.json
161611 build/release/database.json
164173 build/release/table.json
165160 build/release/cql3/statements/select_statement.json
171188 build/release/alternator/executor.json
172547 build/release/mutation_partition.json
200873 build/release/api/storage_service.json
232536 build/release/service/storage_proxy.json
261946 build/release/sstables/sstables.json
275248 build/release/message/messaging_service.json
326384 build/release/repair/repair.json
826227 build/release/repair/row_level.json
```

Next steps
---

1. Split

- repair/repair.cc
- message/messaging_service.cc
- sstables/sstables.cc
- api/storage_service.cc

2. Evaluate the template that takes most time to instantiate globally and that can be instantiated only once